Extract review bundle packet helpers
