+++ scripts/tickets_dispatch_gate.py
@@ -34,6 +34,7 @@ if __package__:
     from .tickets_gate_mutations import _canonical_gate_mutation_plan
     from .tickets_input_producers import GIT_PACKS, git_head, input_groups, render_ticket_inputs
     from .tickets_issue import GATE_ID_MARKER, NEW_DEFAULT_BOUND, _distinct_gate_lenses, _render_ticket
+    from .tickets_gate_bundle import _gate_complete_coverage, _ordered_bundle_carrier, _ordered_lens_bundle
     from .tickets_packet import GATE_CRITIQUE_ID, GATE_EXECUTORS, GATE_EXECUTOR_SECTIONS, GATE_REPAIR_ID, GATE_VERIFY_ID
     from .tickets_store import NO_SINK_ERROR, _create_text_exclusively, _run_lock, _segment_error, _tickets_root
     from .tickets_worklog import _run_tickets
@@ -47,6 +48,7 @@ else:
     from tickets_gate_mutations import _canonical_gate_mutation_plan
     from tickets_input_producers import GIT_PACKS, git_head, input_groups, render_ticket_inputs
     from tickets_issue import GATE_ID_MARKER, NEW_DEFAULT_BOUND, _distinct_gate_lenses, _render_ticket
+    from tickets_gate_bundle import _gate_complete_coverage, _ordered_bundle_carrier, _ordered_lens_bundle
     from tickets_packet import GATE_CRITIQUE_ID, GATE_EXECUTORS, GATE_EXECUTOR_SECTIONS, GATE_REPAIR_ID, GATE_VERIFY_ID
     from tickets_store import NO_SINK_ERROR, _create_text_exclusively, _run_lock, _segment_error, _tickets_root
     from tickets_worklog import _run_tickets
@@ -147,48 +149,6 @@ def _record_names(body: str) -> set:
         if isinstance(name, str):
             names.add(name)
     return names
-def _ordered_bundle_carrier(body: str, lenses: list) -> dict:
-    """The contract-owned root carrier, validated before it is copied."""
-    records = []
-    for group in input_groups(body or ''):
-        if not _is_record(group):
-            continue
-        try:
-            record = parse_canonical_json(group[0][len('- input: '):])
-        except ValueError:
-            continue
-        if isinstance(record, dict):
-            records.append(record)
-    carriers = [record for record in records if record.get('name') == 'ordered-lens-bundle']
-    if len(carriers) != 1:
-        raise ValueError('ordered lens bundle requires the root\'s one canonical `ordered-lens-bundle` Fixed-input record')
-    carrier = carriers[0]
-    rows = carrier.get('value') if carrier.get('type') == 'literal' else None
-    if not isinstance(rows, list) or not rows:
-        raise ValueError('ordered-lens-bundle value must be a non-empty ordered list')
-    identity_inputs = {
-        record.get('name') for record in records if record.get('type') == 'identity'
-    }
-    identities = []
-    used_evidence = set()
-    for position, row in enumerate(rows, start=1):
-        if not isinstance(row, dict) or set(row) != {'evidence', 'identity'}:
-            raise ValueError(f'ordered-lens-bundle entry {position} must contain only identity and evidence')
-        identity = row.get('identity')
-        evidence = row.get('evidence')
-        if not isinstance(identity, str) or not identity.strip() or identity in identities:
-            raise ValueError(f'ordered-lens-bundle entry {position} identity must be non-empty and unique')
-        if not isinstance(evidence, list) or not evidence or any(
-            not isinstance(name, str) or name not in identity_inputs for name in evidence
-        ):
-            raise ValueError(f'ordered-lens-bundle entry {position} evidence must name root identity inputs')
-        if len(evidence) != len(set(evidence)) or used_evidence.intersection(evidence):
-            raise ValueError(f'ordered-lens-bundle entry {position} evidence must be separately attributable')
-        identities.append(identity)
-        used_evidence.update(evidence)
-    if identities != lenses:
-        raise ValueError('ordered-lens-bundle CLI order must equal the root carrier identities unchanged')
-    return carrier
 def _inherited_input_lines(own: str, inherited: str) -> list:
     """The root's fixed-input records the gate does not already state.
     Each is carried as the exact line the root states it on. A gate grades a
@@ -273,18 +233,6 @@ def _gate_input(name: str, *, literal=None, run: str='', ticket: str='', section
     else:
         record = {'name': name, 'type': 'literal', 'value': literal}
     return '- input: ' + json.dumps(record, ensure_ascii=False, separators=(',', ':'), sort_keys=True)
-def _ordered_lens_bundle(value) -> list:
-    lenses = [part.strip() for part in str(value or '').split(',')]
-    if not lenses or any(not lens for lens in lenses):
-        raise ValueError('ordered lens bundle must not contain an empty lens identity')
-    return _distinct_gate_lenses(lenses)
-def _gate_complete_coverage(root: dict, coverage: str) -> bool:
-    rows = []
-    for line in str(coverage or '').splitlines():
-        cells = [cell.strip().strip('`') for cell in line.strip().strip('|').split('|')]
-        if len(cells) >= 2 and cells[0].isdigit(): rows.append((int(cells[0]), cells[1]))
-    criteria = _criteria((root.get('sections') or {}).get('Completion test', ''))
-    return bool(criteria) and rows == [(number, 'gate') for number in range(1, len(criteria) + 1)]
 def _input_name(prefix: str, value: str, position: int) -> str:
     slug = re.sub(r'[^a-z0-9]+', '-', str(value).lower()).strip('-')
     return f'{prefix}-{slug or position}'
